feat: Typed passport getters (#62)
Working with passport data is tricky since the data is untyped. This PR
adds methods to the Passport class to allow typed values to be extracted
where the expected type is known.

Currently, this is only implemented in the oneApp export but it could
also be applied to other locations (i.e. templates). This change should
make it much easier to remove `LooseFlowGraph` and other loose types in
favor of richer/stronger type definitions.

Also in this change is a small interface change where `OneAppPayload`
takes a `sessionId` and an instance of the `Passport` class, removing
the need to pass in a list of files separately. This points in the
direction of starting to use domain objects that encapsulate business
rules rather than repeating the same set of operations on plain data
structures.