chore: Tighten types (#78)
This PR tunes some of the types, removes loose types and resolves all
outstanding linting errors. It also re-introduces a `tsc` check for
`implicitAny`.

This PR makes a few changes which _could_ have changed the output types
and I was a little unsure of some of the export typings in particular. A
thorough review from more seasoned devs would be very much appreciated!

---------

Co-authored-by: Benjamin Williams <benjamin@opensystemslab.io>
Co-authored-by: Dafydd Llŷr Pearson <DafyddLlyr@gmail.com>