refactor: submission email tidy up (#500)
## What does this PR do?

I am tidying up work done on submission email, where it was moved from
the ``teams`` table to the `team_settings`` table

Changes focused on removing old fields and switching from
``NotifyPersonalisation`` in the type to ``TeamContactSettings`` and
using `Pick` to pull them from the ``TeamSettings`` type. This felt
right and in line with other changes.